Change db for PostgreSQL, change way of providing data to app, change model accordingly
Update db to better use of Heroku service, H2 db now is a embedded db for test, DataProvider now uses repository
instead of EntityManager and doesn't crush when there is 503 status on request to API. Even after request with
503 status there should be now possibility for fully using application. One of the best changes is changing
ProductTag and Color to @Embeddable. Product now uses simple @ElementCollection on sets of ProductTags and Colors
instead of more complicated associations.